+++ src/main/kotlin/com/dotori/v2/domain/massage/presentation/dto/req/MassageLimitReqDto.kt
@@ -1,5 +1,5 @@
 package com.dotori.v2.domain.massage.presentation.dto.req
 
 data class MassageLimitReqDto(
-    val limit: Long
+    val limit: Byte
 )
