+++ src/main/kotlin/com/dotori/v2/domain/self_study/presentation/dto/req/SelfStudyLimitReqDto.kt
@@ -1,5 +1,5 @@
 package com.dotori.v2.domain.self_study.presentation.dto.req
 
 data class SelfStudyLimitReqDto(
-    val limit: Long
+    val limit: Short
 )
